[101_runtime_inclusion_dynamic] clarify vars
That is, TT passes all variables currently in scope (including current
loop vars) to the included template, whereas (the method we are using
for runtime inclusion for) TX doesn't.

Thus, we're testing/showing that a top-level variable which is already
in the stash gets "passed on" on TX without explicitly having to pass it
on. We only seem to have to explicitly pass on "local to a loop or maybe
scope" variables.